+++ packages/orm/src/relations/relations.test.ts
@@ -1,7 +1,15 @@
 import { db } from '../test-utils/test-db';
-import { expectSql } from '../test-utils/test-utils';
+import {
+  chatData,
+  expectSql,
+  messageData,
+  userData,
+  useTestDatabase,
+} from '../test-utils/test-utils';
 
 describe('relations', () => {
+  useTestDatabase();
+
   it('should select multiple relations', () => {
     const query = db.user.select({
       profile: (q) => q.profile.where({ bio: 'bio' }),
@@ -33,5 +41,24 @@ describe('relations', () => {
       `,
       ['bio', 1, 'text'],
     );
+  });
+
+  it('should handle sub query pluck', async () => {
+    const chatId = await db.chat.get('id').create(chatData);
+    const authorId = await db.user.get('id').create(userData);
+    await db.message.get('id').create({
+      chatId,
+      authorId,
+      ...messageData,
+    });
+
+    const query = db.user
+      .select({
+        dates: (q) => q.messages.pluck('createdAt'),
+      })
+      .take();
+
+    const result = await query;
+    expect(result).toEqual({ dates: [expect.any(Date)] });
   });
 });