+++ packages/qb/pqb/src/query.ts
@@ -103,15 +103,6 @@ export const queryTypeWithLimitOne = {
 export const isQueryReturnsAll = (q: Query) =>
   !q.query.returnType || q.query.returnType === 'all';
 
-const queryTypeReturningMultipleRows = {
-  undefined: true,
-  all: true,
-  rows: true,
-  pluck: true,
-} as Record<QueryReturnType | 'undefined', true | undefined>;
-export const isQueryReturnsMultipleRows = (q: Query) =>
-  queryTypeReturningMultipleRows[q.query.returnType];
-
 export type JoinedTablesBase = Record<
   string,
   Pick<Query, 'result' | 'table' | 'meta'>